[ADD] Error handling

+++ src/domain/repositories/UserRepository.ts
@@ -17,18 +17,42 @@ export class UserRepository implements IUserRepository {
   }
 
   async getUserById(id: number): Promise<User | undefined> {
-    return await this.userService.getUserById(id);
+    const user = await this.userService.getUserById(id);
+    if (user) {
+      return user;
+    } else {
+      throw new Error(`User with id ${id} not found`);
+    }
   }
 
   async addUser(user: User): Promise<number> {
+    const users = await this.userService.getAllUsers();
+    if (users.some((u) => u.email === user.email)) {
+      throw new Error(`User with email ${user.email} already exists`);
+    }
     return await this.userService.addUser(user);
   }
 
   async updateUser(user: User): Promise<void> {
-    await this.userService.updateUser(user);
+    const findUser = await this.userService.getUserById(user.id);
+    if (findUser) {
+      if (findUser.email !== user.email) {
+        const users = await this.userService.getAllUsers();
+        if (users.some((u) => u.email === user.email)) {
+          throw new Error(`User with email ${user.email} already exists`);
+        }
+      }
+    } else {
+      throw new Error(`User with id ${user.id} not found`);
+    }
   }
 
   async deleteUser(id: number): Promise<void> {
-    await this.userService.deleteUser(id);
+    const user = await this.userService.getUserById(id);
+    if (user) {
+      await this.userService.deleteUser(id);
+    } else {
+      throw new Error(`User with id ${id} not found`);
+    }
   }
 }
